+++ lib/vagrant-openstack/action.rb
@@ -3,7 +3,7 @@ require "pathname"
 require "vagrant/action/builder"
 
 module VagrantPlugins
-  module Rackspace
+  module Openstack
     module Action
       # Include the built-in modules so we can use them as top-level things.
       include Vagrant::Action::Builtin
@@ -18,7 +18,7 @@ module VagrantPlugins
               next
             end
 
-            b2.use ConnectRackspace
+            b2.use ConnectOpenstack
             b2.use DeleteServer
           end
         end
@@ -46,7 +46,7 @@ module VagrantPlugins
       def self.action_read_ssh_info
         Vagrant::Action::Builder.new.tap do |b|
           b.use ConfigValidate
-          b.use ConnectRackspace
+          b.use ConnectOpenstack
           b.use ReadSSHInfo
         end
       end
@@ -57,7 +57,7 @@ module VagrantPlugins
       def self.action_read_state
         Vagrant::Action::Builder.new.tap do |b|
           b.use ConfigValidate
-          b.use ConnectRackspace
+          b.use ConnectOpenstack
           b.use ReadState
         end
       end
@@ -99,7 +99,7 @@ module VagrantPlugins
               next
             end
 
-            b2.use ConnectRackspace
+            b2.use ConnectOpenstack
             b2.use Provision
             b2.use SyncFolders
             b2.use CreateServer
@@ -109,7 +109,7 @@ module VagrantPlugins
 
       # The autoload farm
       action_root = Pathname.new(File.expand_path("../action", __FILE__))
-      autoload :ConnectRackspace, action_root.join("connect_rackspace")
+      autoload :ConnectOpenstack, action_root.join("connect_openstack")
       autoload :CreateServer, action_root.join("create_server")
       autoload :DeleteServer, action_root.join("delete_server")
       autoload :IsCreated, action_root.join("is_created")