+++ lib/vagrant-openstack/action/message_not_created.rb
@@ -1,13 +1,13 @@
 module VagrantPlugins
-  module Rackspace
+  module Openstack
     module Action
-      class MessageAlreadyCreated
+      class MessageNotCreated
         def initialize(app, env)
           @app = app
         end
 
         def call(env)
-          env[:ui].info(I18n.t("vagrant_rackspace.already_created"))
+          env[:ui].info(I18n.t("vagrant_openstack.not_created"))
           @app.call(env)
         end
       end