+++ lib/vagrant-openstack/action/sync_folders.rb
@@ -3,14 +3,14 @@ require 'rbconfig'
 require "vagrant/util/subprocess"
 
 module VagrantPlugins
-  module Rackspace
+  module Openstack
     module Action
       # This middleware uses `rsync` to sync the folders over to the
       # remote instance.
       class SyncFolders
         def initialize(app, env)
           @app    = app
-          @logger = Log4r::Logger.new("vagrant_rackspace::action::sync_folders")
+          @logger = Log4r::Logger.new("vagrant_openstack::action::sync_folders")
           @host_os = RbConfig::CONFIG['host_os']
         end
 
@@ -35,7 +35,7 @@ module VagrantPlugins
               hostpath = hostpath.sub(/^([A-Za-z]):\//, "/cygdrive/#{$1.downcase}/")
             end
 
-            env[:ui].info(I18n.t("vagrant_rackspace.rsync_folder",
+            env[:ui].info(I18n.t("vagrant_openstack.rsync_folder",
                                 :hostpath => hostpath,
                                 :guestpath => guestpath))