+++ lib/vagrant-openstack/errors.rb
@@ -1,0 +1,27 @@
+require "vagrant"
+
+module VagrantPlugins
+  module Openstack
+    module Errors
+      class VagrantOpenstackError < Vagrant::Errors::VagrantError
+        error_namespace("vagrant_openstack.errors")
+      end
+
+      class CreateBadState < VagrantOpenstackError
+        error_key(:create_bad_state)
+      end
+
+      class NoMatchingFlavor < VagrantOpenstackError
+        error_key(:no_matching_flavor)
+      end
+
+      class NoMatchingImage < VagrantOpenstackError
+        error_key(:no_matching_image)
+      end
+
+      class RsyncError < VagrantOpenstackError
+        error_key(:rsync_error)
+      end
+    end
+  end
+end